requirements.txt
anyio==3.7.0
bcrypt==4.0.1
click==8.1.3
colorama==0.4.6
fastapi==0.98.0
h11==0.14.0
idna==3.4
passlib==1.7.4
pydantic==1.10.9
PyJWT==2.7.0
python-multipart==0.0.6
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.0
uvicorn==0.22.0